package rename & put request fix
