+++ resource_management/setup.py
@@ -1,7 +1,7 @@
 from setuptools import find_packages, setup
 
 setup(
-    name="ras_client",
+    name="rasclient",
     version="0.1.0",
     packages=find_packages(),
     include_package_data=True,
