+++ root_orchestrator/resource-abstractor/api/v1/resources_blueprint.py
@@ -7,6 +7,7 @@ from db.clusters_db import (
 )
 from db.clusters_helper import build_filter
 from db.jobs_db import find_job_by_id
+from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint
 from marshmallow import Schema, fields
@@ -46,12 +47,7 @@ class AllResourcesController(MethodView):
         return json_util.dumps(find_clusters(filter))
 
     def put(self, *args, **kwargs):
-        data = None
-        if args and len(args) > 0 and args[0] and type(args[0]) is dict:
-            data = args[0]
-
-        if data is None:
-            raise exceptions.BadRequest()
+        data = request.json
 
         return json_util.dumps(create_cluster(data))
 
@@ -70,12 +66,7 @@ class ResourceController(MethodView):
 
     def patch(self, *args, **kwargs):
         resource_id = kwargs["resourceId"]
-        data = None
-        if args and len(args) > 0 and args[0] and type(args[0]) is dict:
-            data = args[0]
-
-        if data is None:
-            raise exceptions.BadRequest()
+        data = request.json
 
         if ObjectId.is_valid(resource_id) is False:
             raise exceptions.BadRequest()
