+++ root_orchestrator/system-manager-python/blueprints/clusters_blueprints.py
@@ -3,7 +3,7 @@ from ext_requests.cluster_requests import cluster_request_to_delete_job_by_ip
 from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint
-from ras_client import cluster_operations
+from rasclient import cluster_operations
 
 clustersbp = Blueprint("Clusters", "cluster management", url_prefix="/api/clusters")
 clusterinfo = Blueprint("Clusterinfo", "cluster informations", url_prefix="/api/information")
